Move tests from extension to language config file
