+++ src/test/extension.unit.test.ts
@@ -6,16 +6,8 @@
 // tslint:disable:no-any
 
 import { expect } from 'chai';
-import * as sinon from 'sinon';
 import { buildApi } from '../client/api';
 import { EXTENSION_ROOT_DIR } from '../client/common/constants';
-
-// Stub sourceMapSupport.initialize before we import from extension.ts
-// (it's called at the top of the file but we don't need it here)
-// tslint:disable-next-line: no-require-imports no-var-requires
-const sourceMapSupport = require('../client/sourceMapSupport');
-sinon.stub(sourceMapSupport, 'initialize');
-import { DECREASE_INDENT_REGEX, INCREASE_INDENT_REGEX } from '../client/extension';
 
 const expectedPath = `${EXTENSION_ROOT_DIR.fileToCommandArgument()}/pythonFiles/ptvsd_launcher.py`;
 
@@ -29,85 +21,5 @@ suite('Extension API Debugger', () => {
         const args = await buildApi(Promise.resolve()).debug.getRemoteLauncherCommand('something', 1234, true);
         const expectedArgs = [expectedPath, '--default', '--host', 'something', '--port', '1234', '--wait'];
         expect(args).to.be.deep.equal(expectedArgs);
-    });
-    [
-        {
-            keyword: 'def',
-            line: 'def foo:',
-            dedent: false
-        },
-        {
-            keyword: 'class',
-            line: 'class TestClass:',
-            dedent: false
-        },
-        {
-            keyword: 'for',
-            line: 'for item in items:',
-            dedent: false
-        },
-        {
-            keyword: 'if',
-            line: 'if foo is None:',
-            dedent: false
-        },
-        {
-            keyword: 'elif',
-            line: 'elif x < 5:',
-            dedent: true
-        },
-        {
-            keyword: 'else',
-            line: 'else:',
-            dedent: true
-        },
-        {
-            keyword: 'while',
-            line: 'while \'::\' in macaddress:',
-            dedent: false
-        },
-        {
-            keyword: 'try',
-            line: 'try:',
-            dedent: false
-        },
-        {
-            keyword: 'with',
-            line: 'with self.test:',
-            dedent: false
-        },
-        {
-            keyword: 'finally',
-            line: 'finally:',
-            dedent: false
-        },
-        {
-            keyword: 'except',
-            line: 'except TestError:',
-            dedent: false
-        },
-        {
-            keyword: 'async',
-            line: 'async def test(self):',
-            dedent: false
-        }
-    ].forEach(({keyword, line, dedent}) => {
-        test(`Increase indent regex should pick up lines containing the ${keyword} keyword`, async () => {
-            const result = INCREASE_INDENT_REGEX.test(line);
-            expect(result).to.be.equal(true, `Increase indent regex should pick up lines containing the ${keyword} keyword`);
-        });
-
-        test(`Decrease indent regex should ${dedent ? '' : 'not '}pick up lines containing the ${keyword} keyword`, async () => {
-            const result = DECREASE_INDENT_REGEX.test(line);
-            expect(result).to.be.equal(dedent, `Decrease indent regex should ${dedent ? '' : 'not '}pick up lines containing the ${keyword} keyword`);
-        });
-    });
-    test('Increase indent regex should not pick up lines without keywords', async () => {
-        const result = INCREASE_INDENT_REGEX.test('a = \'hello \\n \'');
-        expect(result).to.be.equal(false, 'Increase indent regex should not pick up lines without keywords');
-    });
-    test('Decrease indent regex should not pick up lines without keywords', async () => {
-        const result = DECREASE_INDENT_REGEX.test('a = \'hello \\n \'');
-        expect(result).to.be.equal(false, 'Decrease indent regex should not pick up lines without keywords');
     });
 });
